requirements.txt
django
django-rest-framework
django-storages
django-crontab
django-extensions
django-easy-select2
psycopg2-binary
gunicorn
requests
sh
coverage
django-ckeditor==5.6.1
Pillow==5.4.1
whitenoise==4.1.2
eventlet
python-magic
django-import-export==1.2.0
WeasyPrint==44
mongoengine==0.16.3
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.2.0